split config and implemantation
